Lv01 Refactor
MVC 적용 ) viewController에 있던 내용을 MainView로 구분해 줌으로써
                  책임을 나눔

close : # -
